pyproject.toml
[tool.poetry]
name = "cocksearcher"
version = "0.1.0"
description = ""
authors = ["KIM JUNG BIN <kjbin0420@naver.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
drf-spectacular = "^0.22.0"
django-filter = "^21.1"
serializers = "^0.2.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
